♻️ WVKXG8 prompt: decompose compiler context
