+++ .agentplane/tasks/202605290438-WVKXG8/pr/diffstat.txt
@@ -1,0 +1,3 @@
+ .../src/runtime/prompt-modules/compiler.context.ts | 222 +++++++++++++++++++
+ .../src/runtime/prompt-modules/compiler.ts         | 244 ++-------------------
+ 2 files changed, 238 insertions(+), 228 deletions(-)
