+++ .agentplane/tasks/202605290438-WVKXG8/pr/github-title.txt
@@ -1,0 +1,1 @@
+🚧 WVKXG8 task: Prompt module compiler decomposition [202605290438-WVKXG8]
