+++ .agentplane/tasks/202605290438-WVKXG8/pr/meta.json
@@ -1,0 +1,14 @@
+{
+  "base": "main",
+  "branch": "task/202605290438-WVKXG8/prompt-module-compiler-decomposition",
+  "created_at": "2026-05-29T04:38:43.884Z",
+  "diffstat_sha256": "sha256:8b2c33543cd850cf7e8d7f1f7a969284c92c6f94eab6ef6615ff2c56daea5042",
+  "last_verified_at": "2026-05-29T04:44:24.056Z",
+  "last_verified_diffstat_sha256": "sha256:e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855",
+  "schema_version": 1,
+  "task_id": "202605290438-WVKXG8",
+  "updated_at": "2026-05-29T04:38:43.884Z",
+  "verify": {
+    "status": "pass"
+  }
+}
